backend/requirements.txt
# Core Framework
fastapi==0.110.1
uvicorn==0.25.0
python-multipart==0.0.22
python-dotenv==1.0.1

# Database
motor==3.3.1
pymongo==4.5.0

# Authentication
passlib==1.7.4
bcrypt==4.1.3
PyJWT==2.8.0

# SNMP Monitoring (Python 3.11 compatible)
pyasn1==0.5.1
pyasn1-modules==0.3.0
pysmi-lextudio==1.3.3
pysnmp-lextudio==6.2.0

# MikroTik API
RouterOS-api==0.17.0

# HTTP Client
httpx==0.27.0
requests==2.31.0

# SSH (untuk backup via SSH export)
paramiko==3.4.0


# Validation
pydantic==2.6.0
python-dateutil==2.8.2

# Time-Series Database (InfluxDB 2.x)
influxdb-client==1.50.0
